- move events_home.dart and related widgets to proper package. - delete not required files.
